requirements.txt
basicsr
facexlib
ffmpeg-python
numpy
opencv-python
pillow
psutil
torch
torchvision
tqdm
